requirements.txt
# Flask framework and its core dependencies
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.7.0

# For loading environment variables from a .env file
python-dotenv==1.0.0

# Gunicorn is a production-ready WSGI HTTP server for UNIX
gunicorn==21.2.0
